test(spark): move paimon compact procedure test to Spark 3.5 only
